Move queue to its own module
