+++ src/lib.rs
@@ -4,7 +4,7 @@ pub mod error;
 pub mod event;
 pub mod nats;
 pub mod prelude;
+pub mod queue;
 pub mod storage;
 pub mod vendors;
 pub mod vod;
-pub mod workers;
